inshack-2018/Misc/notrandom/solution/requirements.txt
pwntools
mersenne-twister-predictor
